silentpayments: add shared secret creation routine for sender (a*B)
Note that in contrast to the input public keys tweak, there is no value
in providing a separate routine for determining the input private keys
tweak (i.e. the summed up private keys multiplied with the outpoints
hash, without doing the ECDH with the recipient scan pubkey), as it is
only relevant once for spending.